Moved tests to AdminUpgradeabilityProxy, moved mocks into one file
